EditLinkActivity: Fix #36 Rotating the device clears the fields
It looks like databinding is setting data after the Activity lifecycle
has already restored the saved field values which causes them to be
reset to whatever the Link model values were before the edits.

This is a very hacky approach that uses View.OnAttachStateChangeListener to
attempt to restore the saved bundle values after databinding has reset them
to the Link model. Saving the values to the Link model directly during
onSaveInstanceState might be problematic since the user might not want to save
the draft changes.

Need to look into a cleaner approach, but this should be sufficient since
 EditLinkActivity isn't used much